* origin/topic/vlad/gh-1225:
  Extract length as a uint8
  Change ICMP ND length to a uint16

(cherry picked from commit 5e164469f34ddd894c59eb6304707766f8ae6d4a)

+++ src/analyzer/protocol/icmp/ICMP.cc
@@ -762,7 +762,7 @@ zeek::VectorValPtr ICMP_Analyzer::BuildNDOptionsVal(int caplen, const u_char* da
 			}
 
 		uint8_t type = *((const uint8_t*)data);
-		uint8_t length = *((const uint8_t*)(data + 1));
+		uint16_t length = *((const uint8_t*)(data + 1));
 
 		if ( length == 0 )
 			{
